Refactoring: move components into subdirectories
