+++ src/components/auth/register-form/register-form.jsx
@@ -4,7 +4,7 @@ import { Button, Input } from '@ya.praktikum/react-developer-burger-ui-component
 import { useState } from 'react';
 import { Link, useHistory } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
-import { register } from '../../services/actions/auth';
+import { register } from '../../../services/actions/auth';
