+++ src/components/auth/register-form/register-form.module.css
