+++ src/components/burger/burger-constructor-item/burger-constructor-item.jsx
@@ -2,7 +2,7 @@ import styles from './burger-constructor-item.module.css';
 import { ConstructorElement, DragIcon } from '@ya.praktikum/react-developer-burger-ui-components';
 import { useCallback } from 'react';
 import { useDispatch } from 'react-redux';
-import { ACTION_CONSTRUCTOR_REMOVE, ACTION_CONSTRUCTOR_MOVE } from '../../services/actions/constructor';
+import { ACTION_CONSTRUCTOR_REMOVE, ACTION_CONSTRUCTOR_MOVE } from '../../../services/actions/constructor';
 import { useDrag, useDrop } from 'react-dnd';
 
 function BurgerConstructorItem(item) {
