+++ src/pages/forgot-page.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { ForgotForm } from '../components/forgot-form/forgot-form';
+import { ForgotForm } from '../components/auth/forgot-form/forgot-form';
 
 function ForgotPage() {
