+++ src/pages/login-page.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { LoginForm } from '../components/login-form/login-form';
+import { LoginForm } from '../components/auth/login-form/login-form';
 
 function LoginPage() {
