+++ src/pages/main-page.jsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import { BurgerIngredients } from '../components/burger-ingredients/burger-ingredients';
-import { BurgerConstructor } from '../components/burger-constructor/burger-constructor';
+import { BurgerIngredients } from '../components/burger/burger-ingredients/burger-ingredients';
+import { BurgerConstructor } from '../components/burger/burger-constructor/burger-constructor';
 import { DndProvider } from 'react-dnd';
 import { HTML5Backend } from 'react-dnd-html5-backend';
 import { useSelector } from 'react-redux';
