+++ src/pages/register-page.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { RegisterForm } from '../components/register-form/register-form';
+import { RegisterForm } from '../components/auth/register-form/register-form';
 
 function RegisterPage() {
